profile/requirements.txt
asgiref==3.7.2 
Django==4.2.5 
sqlparse==0.4.4
djangorestframework>=3.14.0
django-filter>=23.3
# psycopg2-binary>=2.9.9
psycopg[binary,pool]
gunicorn>=21.2.0
django-flags